backend/requirements.txt
asgiref==3.3.1
attrs==20.3.0
autopep8==1.5.4
coverage==5.5
Django==3.1.5
djangorestframework==3.12.2
iniconfig==1.1.1
mypy==0.790
mypy-extensions==0.4.3
packaging==20.8
Pillow==8.1.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pyparsing==2.4.7
pytest==6.2.1
pytest-django==4.1.0
pytz==2020.5
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
xkcdpass==1.17.6
